move input logic to common CooperationActionBanner
