+++ src/containers/my-cooperations/accept-cooperation-close/AcceptCooperationClosing.styles.ts
@@ -1,22 +1,5 @@
-import palette from '~/styles/app-theme/app.pallete'
-
 export const styles = {
   boldText: {
     fontWeight: 500
-  },
-  inputBox: {
-    mb: '8px'
-  },
-  inputContainer: {
-    display: 'flex',
-    gap: '16px',
-    height: '50px',
-    mt: '10px'
-  },
-  inputField: {
-    flex: 1
-  },
-  textGray: {
-    color: palette.basic.darkGray
   }
 }
