+++ src/containers/my-cooperations/cooperation-closure-declined-modal/CooperationClosureDeclinedModal.tsx
@@ -1,23 +1,26 @@
-import { FC } from 'react'
+import { useState } from 'react'
 import { ErrorOutlineRounded } from '@mui/icons-material'
 import { useTranslation } from 'react-i18next'
 import { Typography } from '@mui/material'
-import Button from '~/design-system/components/button/Button'
 
+import Button from '~/design-system/components/button/Button'
 import CooperationActionBanner from '~/containers/my-cooperations/cooperation-action-banner/CooperationActionBanner'
 
 import { styles } from './CooperationClosureDeclinedModal.styles'
 
 export interface CooperationClosureDeclinedModalProps {
   message: string
+  onReasonSubmit: (reason: string) => void
   onSend: () => void
   user: string
 }
 
-const CooperationClosureDeclinedModal: FC<
+const CooperationClosureDeclinedModal: React.FC<
   CooperationClosureDeclinedModalProps
-> = ({ message, onSend, user }) => {
+> = ({ message, onReasonSubmit, onSend, user }) => {
   const { t } = useTranslation()
+  const [isInputShown, setIsInputShown] = useState<boolean>(false)
+  const [isReasonSubmitted, setIsReasonSubmitted] = useState<boolean>(false)
 
   return (
     <CooperationActionBanner
@@ -36,6 +39,11 @@ const CooperationClosureDeclinedModal: FC<
         </>
       }
       icon={<ErrorOutlineRounded />}
+      isInputShown={isInputShown}
+      isReasonSubmitted={isReasonSubmitted}
+      onReasonSubmit={onReasonSubmit}
+      setIsInputShown={setIsInputShown}
+      setIsReasonSubmitted={setIsReasonSubmitted}
       title={t('titles.cooperationClosureDeclined')}
     />
   )
